src/requirements.txt
Pillow==8.3.1
pytesseract==0.3.8
PyMuPDF==1.18.15
